Split NIF/Port tests
We can't have a test for both (at least on MacOS), when they are both
in a single Cargo workspace they will share the same dependencies. This
means that when linking the .dylib it will also pick up the other binary
as a dependency and now it's no longer a dynamic library.